requirements.txt
Mastodon.py==1.2.2
